refactor(ui): extract find dialog layout and paint logic — add uint32 settings
Refactored FindItemDialog by extracting layout code to find_item_window_layout.cpp and painting logic to find_item_window_views_paint.cpp, with internal helpers moved to find_item_window_internal.h. Added unsigned integer support to Settings for filter masks, replacing signed integers. Fixed search pagination to use correct map instance and clamp limits properly. Added prefix wildcard matching to item search scoring.
