+++ source/ui/menubar/navigation_menu_handler.cpp
@@ -15,7 +15,7 @@
 
 namespace {
 	[[nodiscard]] uint32_t searchResultsLimit() {
-		return static_cast<uint32_t>(std::max(100000, g_settings.getInteger(Config::SEARCH_RESULTS_LIMIT)));
+		return static_cast<uint32_t>(std::clamp(g_settings.getInteger(Config::SEARCH_RESULTS_LIMIT), 1, 100000));
 	}
 
 	[[nodiscard]] SearchResultRow makeSearchResultRow(uint32_t index, const wxString& name, const Position& position) {
@@ -26,6 +26,20 @@ namespace {
 		};
 	}
 
+	[[nodiscard]] Map* resolvePagedSearchMap(Map* map) {
+		if (map == nullptr) {
+			return nullptr;
+		}
+
+		for (int index = 0; index < g_gui.GetTabCount(); ++index) {
+			if (auto* tab = dynamic_cast<MapTab*>(g_gui.GetTab(index)); tab != nullptr && tab->GetMap() == map) {
+				return tab->GetMap();
+			}
+		}
+
+		return nullptr;
+	}
+
 	void selectFoundBrush(const FindItemDialog& dialog) {
 		if (const Brush* brush = dialog.getResult()) {
 			const PaletteType palette = dialog.getResultKind() == AdvancedFinderCatalogKind::Creature ? TILESET_CREATURE : TILESET_RAW;
@@ -33,11 +47,16 @@ namespace {
 		}
 	}
 
-	void showItemSearchResults(uint16_t item_id, const wxString& load_bar_label, uint32_t page_offset = 0) {
+	void showItemSearchResults(Map* map, uint16_t item_id, const wxString& load_bar_label, uint32_t page_offset = 0) {
+		Map* search_map = resolvePagedSearchMap(map);
+		if (search_map == nullptr) {
+			return;
+		}
+
 		const uint32_t page_limit = searchResultsLimit();
 		EditorOperations::ItemSearcher finder(item_id, page_limit, page_offset);
 		g_gui.CreateLoadBar(load_bar_label);
-		foreach_ItemOnMap(g_gui.GetCurrentMap(), finder, false);
+		foreach_ItemOnMap(*search_map, finder, false);
 		g_gui.DestroyLoadBar();
 
 		SearchResultWindow* window = g_gui.ShowSearchWindow();
@@ -52,17 +71,22 @@ namespace {
 			finder.totalMatches,
 			page_offset,
 			page_limit,
-			[item_id, load_bar_label](uint32_t next_offset) {
-				showItemSearchResults(item_id, load_bar_label, next_offset);
+			[map, item_id, load_bar_label](uint32_t next_offset) {
+				showItemSearchResults(map, item_id, load_bar_label, next_offset);
 			}
 		);
 	}
 
-	void showCreatureSearchResults(const Brush* creature_brush, const wxString& load_bar_label, uint32_t page_offset = 0) {
+	void showCreatureSearchResults(Map* map, const Brush* creature_brush, const wxString& load_bar_label, uint32_t page_offset = 0) {
+		Map* search_map = resolvePagedSearchMap(map);
+		if (search_map == nullptr) {
+			return;
+		}
+
 		const uint32_t page_limit = searchResultsLimit();
 		EditorOperations::CreatureSearcher finder(creature_brush, page_limit, page_offset);
 		g_gui.CreateLoadBar(load_bar_label);
-		foreach_TileOnMap(g_gui.GetCurrentMap(), finder);
+		foreach_TileOnMap(*search_map, finder);
 		g_gui.DestroyLoadBar();
 
 		SearchResultWindow* window = g_gui.ShowSearchWindow();
@@ -77,8 +101,8 @@ namespace {
 			finder.totalMatches,
 			page_offset,
 			page_limit,
-			[creature_brush, load_bar_label](uint32_t next_offset) {
-				showCreatureSearchResults(creature_brush, load_bar_label, next_offset);
+			[map, creature_brush, load_bar_label](uint32_t next_offset) {
+				showCreatureSearchResults(map, creature_brush, load_bar_label, next_offset);
 			}
 		);
 	}
@@ -151,10 +175,15 @@ void NavigationMenuHandler::OnJumpToItemBrush(wxCommandEvent& WXUNUSED(event)) {
 	const int modal_result = dialog.ShowModal();
 	if (modal_result != wxID_CANCEL) {
 		if (dialog.getResultAction() == FindItemDialog::ResultAction::SearchMap) {
+			if (!g_gui.IsEditorOpen()) {
+				return;
+			}
+
+			Map* current_map = &g_gui.GetCurrentMap();
 			if (dialog.getResultKind() == AdvancedFinderCatalogKind::Creature) {
-				showCreatureSearchResults(dialog.getResult(), "Searching map...");
+				showCreatureSearchResults(current_map, dialog.getResult(), "Searching map...");
 			} else {
-				showItemSearchResults(dialog.getResultID(), "Searching map...");
+				showItemSearchResults(current_map, dialog.getResultID(), "Searching map...");
 			}
 		} else if (dialog.getResultAction() == FindItemDialog::ResultAction::SelectItem) {
 			selectFoundBrush(dialog);
